+++ packages/web/src/lib/view/parts-list/filter/FilterPanel.svelte
